Add //mojo to fuchsia
This mojo repository will eventually contain the Mojo public headers and an
implementation of those headers for Fuchsia. It's possible it will end up
containing all of https://github.com/domokit/mojo or that we'll end up
splitting https://github.com/domokit/mojo into multiple repos as we integrate
it with Fuchsia.

Change-Id: I5e81e974419d248af83836dabd7414a98b801b57