Pin Dart @de63e4fa94a1 to match flutter/engine
This pins the version of Dart (specifically github.com/dart-lang/sdk)
to the same version used by flutter/engine and documents the process
for updating to a newer version. This is to insulate the Fuchsia build
from the relatively frequent breakages in the Dart build and keep
the Flutter engine build in Fuchsia consistent.

The Dart revision in Fuchsia should generally be updated whenever the
version of Dart used in Flutter is updated, although not all rolls to
Flutter will be have meaningful changes for Fuchsia.

Change-Id: Id664f43266120e949dedda5919be582f6b7d57be