Remove llvm repositories from runtimes/cpp
Now that we get libc++/etc from the toolchain repository, we don't need
to have the llvm projects checked out on every Fuchsia developer's
machines in order to build Fuchsia. Developers working on the toolchain
themselves can use the 'toolchain' manifest to get access to these
bits. This shaves ~600mb off of the checkout size.

Change-Id: Ia84e678b4af668db90621cd699ea39a77f3ec392