Remove llvm repositories from runtimes/cpp (which removes the manifest)
Now that we get libc++/etc from the toolchain repository, we don't need
to have the llvm projects checked out on every Fuchsia developer's
machines in order to build Fuchsia. Developers working on the toolchain
themselves can use the 'toolchain' manifest to get access to these
bits. This shaves ~600mb off of the checkout size.

Since this is the last thing in the manifest, also remove the whole
thing. The only things needed for building C++ code are in the default
set (i.e. magenta + buildtools).

Change-Id: Icb113dc8d7a7e0235c1ba605bcfc26df87d47b93