Add runtimes/go_test
This is lighter than the fuchsia manifest, and it only includes what's
necessary to build with "--modules go_stdlib_tests,boot_test_runner".
However, it is brittle, due to the possibility of new transitive GN
dependencies being introduced. When that happens, the build will fail
with a message about missing files.

The dependencies on Dart and Rust are because Test Runner includes a
FIDL service, which gets compiled for all languages.

Change-Id: Id868435f19567540ef7e5cd3666fba174e88dccc