Temporarily add repos required by fidl to drivers manifest
Now that the wifi drivers have a FIDL dependency, we need to pull
additional repositories that are required by the current FIDL gn rules.
These are:

- lib/fidl to actually generate the code.

- lib/ftl as a dependency of the generated C++ code.

- runtimes/rust and runtimes/dart are needed by the rules generated to
produce Rust and Dart code from FIDL interfaces. There are no consumers
of these bindings in this configuration, but our build system isn't
smart enough to know that now so it blindly generates rules for them
anyway.

- third_party/jinja2 and third_party/markupsafe which are used by the
current python-based generation backends.

The need to download additional language runtimes is a really
unfortunate consequence of the way our code generation targets expand.
The basic issue is that the generation rule doesn't know what languages
it needs to generate for. A better fix would be to configure the
fidl template to only expand out the ones that are needed.

Change-Id: Ib99c55f4c8730ec7c5e14e26933eb6258242359d