Move Xi and rust_ledger_example to userspace
These two projects aren't actually required as part of the Rust runtime
to build other Rust projects, they are just projects written in Rust.
Thus they seem to belong in the userspace manifest file and not
runtimes/rust.

It's plausible rust_ledger_example should belong in the ledger manifest
but then that would depend on rust, which it doesn't currently. And the
userspace manifest already depends on rust and also seems reasonable.

Change-Id: I954a5fcc013dbc22edbce96b95082c4d7b195c4e