Adds manifest for googletest.
Adds a dependency on googletest to cobalt.

The drivers project is also importing googletest but into a
sub-directory of third_paryt/shaaderc for some reason. I did
not want to mess with that. But it was necessary to change the
name of the project or Jiri complained. I also fixed the formating.

Change-Id: Ibab26503734d1546a38877c769a36166676571b6