Cobalt manifest: switch |remotebranch| to |revision|.
After working with it for a while we have decided that having
a special branch for Fuchsia is not really what we want.
Pinning a revision is a better solution.

Among other reasons I perfer this: There is exactly one
place that controls which Cobalt code is included in the
Fuchsia build: the file manifest/cobalt. The problem with
the |remotebranch| solution is that a remote branch is
only a pointer and the referent of that pointer may be
changed without any evidence of that in the Fuchsia repo.

Change-Id: Ieca34f29631b4a69568eac6fb150c1ceb0dd3d8e