userspace manifest now depends on cobalt_client.
This is being done for two reasons:
(1) To make the cobalt_client build in LUCI work we are switching
it to depend on the Fuchsia manifest and then pulling in the
cobalt manifest through this userspace manifest.

(2) We want userspace to start depending on cobalt_client anyway
because we are now introducing cobalt_client as a general
userspace service.

Change-Id: I58009fca0507d5465a7a8118fd29111904f64f95