Pin revisions of all repositories in toolchain manifest
Goma expects version for the Linux and Mac toolchain to be the
same. However, since Linux and Mac bots don't start building at
exactly the same moment, and since our toolchain repositories are
now automatically synchronized with upstream, we don't have
a guarantee that all the repositories will be at the same revision.
Pinning all repositories to revision enforces that property.

We don't yet have a way to automatically roll the revisions (TO-404).

Change-Id: Iaa3374724c52b918764b3d669998182afd40a5dc