[go] Pin //third_party/go to fixed revision temporarily
This pins the code at //third_party/go to revision 07a4bc25 in order to
stage a standard library change that has to appear simultaneously in
our prebuilt Go toolchain and the source tree since they change the
layout of packages. I'll unpin when rolling the Go prebuilts.

This also adds a 'go_head' manifest so the bot that makes the prebuilts
can use versions newer than the one we pin Fuchsia to.

Change-Id: I3fc4434316f96c15494ea235dee821157e630bf5